% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_dom_channel.R
\name{dom_channel}
\alias{dom_channel}
\title{Dominant Channel}
\usage{
dom_channel(wav_file)
}
\arguments{
\item{wav_file}{the original \code{wav_file} path}
}
\description{
Finds the dominant channel for part of the conversation to determine which
speaker is speaking.
}
