% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_make_textgrid.R
\name{make_textgrid}
\alias{make_textgrid}
\title{Make TextGrid}
\usage{
make_textgrid(data, wav_file)
}
\arguments{
\item{data}{The finalized data from the \code{clean_up()} function}

\item{wav_file}{The path to the wav file}
}
\description{
Takes all of the results and formats it into a TextGrid.
}
