% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_remove_noises.R
\name{remove_noises}
\alias{remove_noises}
\title{remove_noises}
\usage{
remove_noises(ch1, ch2)
}
\arguments{
\item{ch1}{channel 1 path}

\item{ch2}{channel 2 path}
}
\description{
Removes some of the extra noise in each channel
}
