% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_textgrid.R
\name{auto_textgrid}
\alias{auto_textgrid}
\title{Automate TextGrid}
\usage{
auto_textgrid(wav_file)
}
\arguments{
\item{wav_file}{The path to the wav file}
}
\description{
Takes all of the steps and puts them together to produce a textgrid from a
wav file.
}
