% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_textgrid.R
\name{auto_textgrid}
\alias{auto_textgrid}
\title{Automate TextGrid}
\usage{
auto_textgrid(
  wav_file,
  noise_reduction = FALSE,
  min_pitch = 100,
  time_step = 0,
  threshold = -45,
  min_silent_int = 0.5,
  min_sound_int = 0.1,
  model_type = "base",
  prompt = NULL,
  plot = FALSE,
  remove_partial = FALSE,
  hyphen = "keep",
  remove_apostrophe = FALSE,
  remove_punct = FALSE,
  whisp = NULL
)
}
\arguments{
\item{wav_file}{The path to the wav file}

\item{noise_reduction}{whether the praat noise reduction script should be run before getting boundaries, default = FALSE}

\item{min_pitch}{Minimum pitch (Hz)}

\item{time_step}{Time step (s)}

\item{threshold}{silence threshold, default is -45}

\item{min_silent_int}{Minimum silent interval (s)}

\item{min_sound_int}{Minimum sounding interval (s)}

\item{model_type}{the Whisper model type to use, default is "base".
Other options are in order of complexity: tiny, base, small, medium, and large
(see https://github.com/openai/whisper/blob/main/model-card.md).}

\item{prompt}{Can prompt the model with words, names, spellings you want it to use.
Default prompts the use of backchannels, repetitions, and other conversational patterns.}

\item{plot}{plot the amplitude? Default is FALSE.}

\item{remove_partial}{Should the model keep words that are incomplete (e.g., "I don- I don't know")? Default is FALSE.}

\item{hyphen}{Should hyphens be retained or replaced? Options are "space" (hyphens are replaced with a space), "keep" (the hyphens are retained), "remove" the hyphens are removed with no white space added.}

\item{remove_apostrophe}{Should all apostraphes be removed? Default is FALSE.}

\item{remove_punct}{Should all punctuation be removed (other than hyphens and apostrophes)? Default is FALSE.}

\item{whisp}{the reticulated whisper model (e.g. produced via \verb{whisper = reticulate::import("whisper"); model = whisper$load_model(model_type)})}
}
\description{
Takes all of the steps and puts them together to produce a textgrid from a
wav file.
}
