% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_textgrid.R
\name{auto_textgrid}
\alias{auto_textgrid}
\title{Automate TextGrid}
\usage{
auto_textgrid(
  wav_file,
  noise_reduction = FALSE,
  min_pitch = 100,
  time_step = 0,
  threshold = -45,
  min_silent_int = 0.5,
  min_sound_int = 0.1,
  model_type = "base"
)
}
\arguments{
\item{wav_file}{The path to the wav file}

\item{noise_reduction}{whether the praat noise reduction script should be run before getting boundaries, default = FALSE}

\item{min_pitch}{Minimum pitch (Hz)}

\item{time_step}{Time step (s)}

\item{threshold}{silence threshold, default is -45}

\item{min_silent_int}{Minimum silent interval (s)}

\item{min_sound_int}{Minimum sounding interval (s)}

\item{model_type}{the Whisper model type to use, default is "base". Other options are in order of complexity: tiny, base, small, medium, and large (see https://github.com/openai/whisper/blob/main/model-card.md).}
}
\description{
Takes all of the steps and puts them together to produce a textgrid from a
wav file.
}
