% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clean_up.R
\name{clean_up}
\alias{clean_up}
\title{Clean Up Whisper Output}
\usage{
clean_up(whispered1, whispered2, folder)
}
\arguments{
\item{whispered1}{channel one whisper data output}

\item{whispered2}{channel two whisper data output}

\item{folder}{the folder where the files are located}
}
\description{
Cleans up the data to get it ready for making
the TextGrid file.
}
