% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_clean_up.R
\name{clean_up}
\alias{clean_up}
\title{Clean Up Results}
\usage{
clean_up(result_full, result1, doms)
}
\arguments{
\item{result_full}{The Whisper results for the full conversation}

\item{result1}{The Whisper results for the channel 1 of the conversation}

\item{doms}{The results of the dominant channel analysis}
}
\description{
Cleans up the data produced by Whisper and the dominant channel analysis.
}
