% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Get Boundaries}
\usage{
get_boundaries(folder, threshold = -45)
}
\arguments{
\item{folder}{the folder where the wav_file is located}

\item{threshold}{silence threshold, default is -45}
}
\description{
Uses Praat to find boundaries of speaking turns.
}
