% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_make_textgrid.R
\name{make_textgrid}
\alias{make_textgrid}
\title{Make TextGrid}
\usage{
make_textgrid(data, wav_file)
}
\arguments{
\item{data}{cleaned data from clean_up().}

\item{wav_file}{The wave file}
}
\description{
Cleans up the data produced by Whisper and the dominant channel analysis.
}
