% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_py_setup.R
\name{py_setup}
\alias{py_setup}
\title{Python Set Up}
\usage{
py_setup(path, install = TRUE)
}
\arguments{
\item{path}{The path to the miniconda environment}

\item{install}{Whether to install conda at the specified path (default = TRUE)}
}
\description{
The function sets up a miniconda environment at \code{path_to_env}
}
