% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_py_setup.R
\name{py_setup}
\alias{py_setup}
\title{Python Set Up}
\usage{
py_setup(path_to_env)
}
\arguments{
\item{path_to_env}{The path to the miniconda environment}
}
\description{
The function sets up a miniconda environment at \code{path_to_env}
}
