% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get_boundaries.R
\name{set_praat_path}
\alias{set_praat_path}
\title{Praat File Path}
\usage{
set_praat_path(path = NULL)
}
\arguments{
\item{path}{The path to a Praat executable file. If NULL, then uses the default for the system.}
}
\description{
This allows the user to define where their Praat installation is located if not in the default.
}
