% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_split_channels.R
\name{split_channels}
\alias{split_channels}
\title{split_channels}
\usage{
split_channels(wav_file)
}
\arguments{
\item{wav_file}{The path to the wav file}
}
\description{
Splits the wav file into the two channels
}
