% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_whispering.R
\name{whispering}
\alias{whispering}
\title{Whisper Transcription}
\usage{
whispering(ch1, ch2 = NULL, folder, model_type, prompt, whisp = NULL)
}
\arguments{
\item{ch1}{channel 1 file}

\item{ch2}{channel 2 file (NULL for single channel wav files)}

\item{folder}{folder of the files}

\item{model_type}{the type of Whisper model to run}

\item{prompt}{Can prompt the model with words, names, spellings you want it to use.}

\item{whisp}{the reticulated whisper model (e.g. produced via \verb{whisper = reticulate::import("whisper"); model = whisper$load_model(model_type)})}
}
\description{
Applies the OpenAI Whisper model to transcribe the conversation
}
