% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_whispering.R
\name{whispering}
\alias{whispering}
\title{Whisper Transcription}
\usage{
whispering(wav_file, final_ch1, model_type = "base")
}
\arguments{
\item{wav_file}{the original \code{wav_file} path}

\item{final_ch1}{channel 1 path after the cleaning from \code{remove_noises()}}

\item{model_type}{the type of Whisper model to run, default is "base"}
}
\description{
Applies the OpenAI Whisper model to transcribe the conversation
}
