% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_whispering.R
\name{whispering}
\alias{whispering}
\title{Whisper Transcription}
\usage{
whispering(ch1, ch2, folder, model_type = "base")
}
\arguments{
\item{ch1}{channel 1 file}

\item{ch2}{channel 2 file}

\item{folder}{folder of the files}

\item{model_type}{the type of Whisper model to run, default is "base"}
}
\description{
Applies the OpenAI Whisper model to transcribe the conversation
}
